Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -5,7 +5,7 @@ int main(){
     int c;
     scanf("%d %d %d",&a ,&b ,&c);
     float avg ;
-    avg = (a+b+c)/3;
+    avg = (a+b+c)/(float)3;
     printf("Average: %.2f",avg);
     return 0;
 }
